portfolio_django_admin/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.6.0",
    "asgiref==3.8.1",
    "async-property==0.2.2",
    "attrs==24.2.0",
    "autobahn==24.4.2",
    "automat==24.8.1",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "constantly==23.10.4",
    "cryptography==43.0.1",
    "daphne==4.1.2",
    "debugpy==1.8.5",
    "deprecation==2.1.0",
    "distro==1.9.0",
    "django==5.1",
    "django-appconf==1.0.6",
    "django-ckeditor==6.7.1",
    "django-compressor==4.5.1",
    "django-cors-headers==4.6.0",
    "django-js-asset==2.2.0",
    "django-libsass==0.9",
    "django-ninja==1.4.3",
    "django-o365mail==1.1.0",
    "django-redis==5.4.0",
    "django-sass-processor==1.4.1",
    "ecdsa==0.19.0",
    "et-xmlfile==1.1.0",
    "graphql-core==3.2.5",
    "graphql-relay==3.2.0",
    "h11==0.14.0",
    "hiredis==3.0.0",
    "httpcore==1.0.5",
    "httpx==0.27.2",
    "hyperlink==21.0.0",
    "idna==3.10",
    "incremental==24.7.2",
    "jiter==0.11.1",
    "josepy==1.14.0",
    "jwcrypto==1.5.6",
    "libcst==1.5.0",
    "libsass==0.23.0",
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    "mozilla-django-oidc==4.0.1",
    "numpy==2.1.1",
    "o365==2.0.36",
    "oauthlib==3.2.2",
    "openpyxl==3.1.5",
    "packaging==24.1",
    "pandas==2.2.2",
    "promise==2.3",
    "psycopg2-binary==2.9.9",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pygments==2.18.0",
    "pyjwt==2.10.1",
    "pyopenssl==24.2.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-jose==3.3.0",
    "python-keycloak==4.4.0",
    "python-multipart==0.0.16",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "rcssmin==1.1.2",
    "redis==5.2.0",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==13.9.3",
    "rjsmin==1.2.2",
    "rsa==4.9",
    "service-identity==24.2.0",
    "setuptools==72.1.0",
    "shellingham==1.5.4",
    "six==1.16.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlparse==0.5.1",
    "starlette==0.41.2",
    "strawberry-graphql==0.247.0",
    "text-unidecode==1.3",
    "twisted==24.10.0",
    "txaio==23.1.1",
    "typer==0.12.5",
    "typing-extensions==4.12.2",
    "tzdata==2024.1",
    "tzlocal==5.2",
    "urllib3==2.2.3",
    "uvicorn==0.32.0",
    "wheel==0.44.0",
    "zope-interface==7.1.1",
]
